4/17

+++ database/migrations/2024_03_31_223030_create_offers_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->id();
             $table->string('name');
             $table->string('premalink')->unique();
-            $table->integer('start_date');
-            $table->integer('end_date');
+            $table->bigInteger('start_date');
+            $table->bigInteger('end_date');
             $table->string('cashback_amount');
             $table->integer('thumbnail')->nullable();
             $table->bigInteger('shop_id')->unsigned()->index();
